matrix 559
Raman is very good in programming. His teacher wants to test Raman’s skills and gives him a problem to code. He is given a matrix pattern and is asked to write a code to print the same. The pattern is different for different values of input, but the logic for all the patterns is the same. Help Raman to develop the code for the same, by referring to the sample input and output.

 

Input Description:
A positive integer N

Output Description:
A particular matrix pattern

Sample Input :
5

Sample Output :
5 4 3 2 1 2 3 4 5
4 4 3 2 1 2 3 4 4
3 3 3 2 1 2 3 3 3
2 2 2 2 1 2 2 2 2
1 1 1 1 1 1 1 1 1
2 2 2 2 1 2 2 2 2
3 3 3 2 1 2 3 3 3
4 4 3 2 1 2 3 4 4
5 4 3 2 1 2 3 4 5