matrix 426
You are given a configuration of a 4 x 4 chess board. An empty square is represented by a ‘.’ and an occupied square is represented by a ‘#’. You are given Q queries of the form “x y” denoting the coordinate of the square (one of (0 0) to (3 3)). If a knight can be placed on that square, print the number of squares it can reach from that square, print -1 if the square is blocked and the knight cannot be placed. A blocked square cannot be reached.
For Reference: https://en.wikipedia.org/wiki/Knight_(chess)#Movement

Input Description:
You are given a 4 by 4 character matrix denoting the board configuration The next line contains the number of queries Q The next Q lines contain a coordinate pair.

Output Description:
Print Q lines answering each query

Sample Input :
....
..#.
....
....
2
0 0
1 2

Sample Output :
1
-1